Simplifying hosted stuff ✨ (MigrationTools.Host): add base command and execute migration command
Introduce `CommandBase` class to handle common command logic, including
startup logic, telemetry, and version checks. Add `ExecuteMigrationCommand`
to handle the execution of migration tasks, utilizing the `CommandBase`
class for shared functionality. Also, add `ExecuteMigrationCommandSettings`
to define the settings required for the migration command.

These changes are made to modularize the command logic, making it easier
to manage and extend. The `CommandBase` class centralizes common tasks
like telemetry and version checks, reducing code duplication. The
`ExecuteMigrationCommand` class focuses on the specific logic needed to
execute a migration, improving code readability and maintainability.

✨ (InitMigrationCommand): add InitMigrationCommand to initialize migration configurations

✨ (InitMigrationCommandSettings): add settings for InitMigrationCommand

🔧 (MigrationToolHost): update command registration for InitMigrationCommand

The InitMigrationCommand is introduced to allow users to initialize migration configurations with various options. This command supports different configuration modes such as Reference, Basic, WorkItemTracking, Fullv2, and WorkItemTrackingv2, making it flexible for different migration scenarios.

The InitMigrationCommandSettings class is added to handle the settings for the InitMigrationCommand, including the output mode which determines the type of configuration to be generated.

The MigrationToolHost is updated to register the new InitMigrationCommand, ensuring it is available for use within the application. This enhances the tool's functionality by providing a straightforward way to set up migration configurations.

♻️ (StartupService.cs): remove version check and telemetry logging for simplification

Remove the extensive version check and telemetry logging to simplify the startup process. This reduces the complexity and potential overhead during application startup, making the codebase easier to maintain and understand. The removed code includes checks for package manager installation, version updates, and telemetry notes, which are deemed unnecessary for the current scope of the application.
